+++ CHANGELOG.md
@@ -4,6 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2021-11-02
+### Added
+- utility class \MageGyver\M2devbox\Util\Version
+- support for Magento 2.4.2-p2, 2.4.3 and 2.4.3-p1
+
+### Changed
+- updated name banner in help message
+
+### Fixed
+- using wrong autoload.php when m2devbox was installed as a Composer package
+
 ## [1.1.0] - 2021-06-21
 ### Added
 - m2devbox now includes a Redis container and uses it for Magento caching and page caching
@@ -58,7 +69,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - initial public release
 
-[Unreleased]: https://github.com/MageGyver/m2devbox/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/MageGyver/m2devbox/compare/1.2.0...HEAD
+[1.2.0]: https://github.com/MageGyver/m2devbox/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/MageGyver/m2devbox/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/MageGyver/m2devbox/compare/0.2.2...1.0.0
 [0.2.2]: https://github.com/MageGyver/m2devbox/compare/0.2.1...0.2.2